requirements.txt
psutil
Twisted
requests
fuzzywuzzy
beautifulsoup4
python-dateutil
pymysql
typing
